The City and County of Denver 2019 Procurement Transactions financial dataset
Donut chart in power Bi, presented the most understandable analysis for the program area which received the most amount and that was airport, which consisted of about 749.39M of payments in the year 2019.  The donut chart in Power Bi helped to create a representation of different program area with respect to the total funding received in each area, the legends on the side of the chart helps the reader to understand the different program area assigning a different color and size with respect to the amount of payments made in the donut chart, the numbers on the side of each color and program area gives us the total amount of payments made in the year 2019 for the specific program area, making it a well understandable representation. 

Funnel chart made with the help of Power Bi, gave us the result for the program area which received the lease amount of payment, that is economic opportunity. Funnel chart in power BI gave us a very appropriate representation of the visual which was in a shape of a funnel with the largest area representing the largest payment and the smallest area the least. The legend on the right of the chart made gave us a correct representation of different program areas. 

Tree map made with the help of power Bi gave us the state in the airport department which received the largest amount of payment and that is CO. Tree map in power Bi helped us analyze multiple values, the amount received in each program area and amount received in each state in that program area. This analysis in tree map gave it an advantage over other maps used for the same analysis.